+++ server/component/style/validate/tpl_validate_anonymous_user.php
@@ -1,0 +1,53 @@
+<?php
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+?>
+<div class="container mt-3 <?php echo $this->css; ?>">
+    <?php $this->output_alert(); ?>
+    <div class="jumbotron">
+        <h1><?php echo $this->title; ?></h1>
+    </div>
+    <div class="card mb-3">
+        <div class="card-header">
+            <h5 class="mb-0"><?php echo $this->subtitle; ?></h5>
+        </div>
+        <div class="card-body">
+            <form action="<?php echo $_SERVER['REQUEST_URI']; ?>" method="post">
+                <div class="form-group d-none">
+                    <label>Leave this field empty</label>
+                    <input type="text" class="form-control" name="phone7h92jP" autocomplete="off">
+                </div>
+                <div class="form-group <?php echo $this->get_css_name(); ?>">
+                    <label><?php echo $this->name_label; ?></label>
+                    <input disabled type="text" class="form-control" autocomplete="username" name="name" placeholder="<?php echo $this->name_placeholder; ?>" value="<?php echo $name; ?>" required>
+                    <small class="form-text text-muted"><?php echo $this->anonymous_user_name_description; ?></small>
+                </div>
+                <div class="form-group">
+                    <label><?php echo $this->pw_label; ?></label>
+                    <input type="password" autocomplete="new-password" class="form-control mb-1" name="pw" placeholder="<?php echo $this->pw_placeholder; ?>" required>
+                    <input type="password" autocomplete="new-password" class="form-control" name="pw_verify" placeholder="<?php echo $this->pw_confirm_label; ?>" required>
+                </div>
+                <div class="form-group <?php echo $this->get_css_gender(); ?>">
+                    <div>
+                    <label><?php echo $this->gender_label; ?></label>
+                    </div>
+                    <div class="form-check form-check-inline">
+                    <input class="form-check-input" type="radio" name="gender" value="1" <?php echo $male_checked; ?> required>
+                        <label class="form-check-label"><?php echo $this->gender_male; ?></label>
+                    </div>
+                    <div class="form-check form-check-inline">
+                        <input class="form-check-input" type="radio" name="gender" value="2" <?php echo $female_checked; ?> required>
+                        <label class="form-check-label"><?php echo $this->gender_female; ?></label>
+                    </div>
+                    <div class="form-check form-check-inline">
+                        <input class="form-check-input" type="radio" name="gender" value="3" <?php echo $divers_checked; ?> required>
+                        <label class="form-check-label"><?php echo $this->gender_divers; ?></label>
+                    </div>
+                </div>
+                <?php $this->check_custom_fields(); ?>
+                <button type="submit" class="btn btn-primary"><?php echo $this->activate_label; ?></button>
+            </form>
+        </div>
+    </div>
+</div>